# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "老年人用药管理系统",
    "description": "老人用药管理智能助手后端 API",
    "version": "2.33.0"
  },
  "paths": {
...
    "/api/v1/auth/webauthn/register/options": {
      "post": {
        "tags": [
          "auth",
          "webauthn"
        ],
        "summary": "Wa Register Options",
        "description": "构造通行密钥登记选项（已排除本用户已有凭证）。",
        "operationId": "wa_register_options_api_v1_auth_webauthn_register_options_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebAuthnOptionsOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/webauthn/register": {
      "post": {
        "tags": [
          "auth",
          "webauthn"
        ],
        "summary": "Wa Register",
        "description": "校验登记响应并保存凭证。",
        "operationId": "wa_register_api_v1_auth_webauthn_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebAuthnRegisterIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/auth/webauthn/login": {
      "post": {
        "tags": [
          "auth",
          "webauthn"
        ],
        "summary": "Wa Login",
        "description": "校验通行密钥断言，按 credential_id 反查用户并签发 JWT。",
        "operationId": "wa_login_api_v1_auth_webauthn_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebAuthnLoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResp"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/webauthn/credentials": {
      "get": {
        "tags": [
          "auth",
          "webauthn"
        ],
        "summary": "Wa List Credentials",
        "description": "列出当前用户已登记的通行密钥。",
        "operationId": "wa_list_credentials_api_v1_auth_webauthn_credentials_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/webauthn/credentials/{cred_id}": {
      "delete": {
        "tags": [
          "auth",
          "webauthn"
        ],
        "summary": "Wa Delete Credential",
        "description": "删除指定通行密钥。",
        "operationId": "wa_delete_credential_api_v1_auth_webauthn_credentials__cred_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cred_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cred Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "用户"
        ],
        "summary": "Read Current User",
        "description": "获取当前用户信息",
        "operationId": "read_current_user_api_v1_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "用户"
        ],
        "summary": "Update Current User",
        "description": "更新当前用户信息",
        "operationId": "update_current_user_api_v1_users_me_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "用户"
        ],
        "summary": "Delete Current User",
        "description": "注销当前用户账号（硬删除，会级联删除用药计划/记录/AI日志）",
        "operationId": "delete_current_user_api_v1_users_me_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/bind": {
      "post": {
        "tags": [
          "用户"
        ],
        "summary": "Bind Family",
        "description": "家属绑定老人（组成家庭组，并把 device_id 关联到真实老人）\n\n绑定逻辑：\n1. 家属身份校验\n2. 校验目标用户是真实老人\n3. 校验 device_id 对应的虚拟用户存在（即设备已通过 device/register 注册）\n   - 若虚拟用户不存在，说明设备未注册或 device_id 错误\n   - 若虚拟用户存在，绑定后会被迁移数据并删除，device_id 关联到真实老人\n4. 调用 UserService.bind_family 完成数据迁移 + device_id 关联 + 家庭组绑定\n\n:param req: 绑定请求（elderly_user_id + device_id）\n:param db: 数据库会话\n:param current_user: 当前登录用户（必须是家属）\n:return: {\"group_id\": int}",
        "operationId": "bind_family_api_v1_users_bind_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindFamilyReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/{user_id}": {
      "delete": {
        "tags": [
          "用户"
        ],
        "summary": "Delete User By Family",
        "description": "家属删除同家庭组的老人账号（硬删除）",
        "operationId": "delete_user_by_family_api_v1_users__user_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BindFamilyReq": {
        "properties": {
          "elderly_user_id": {
            "type": "integer",
            "title": "Elderly User Id",
            "description": "老人用户ID"
          },
          "device_id": {
            "type": "string",
            "title": "Device Id",
            "description": "老人设备ID（弱保护：需知道设备ID才能绑定）"
          }
        },
        "type": "object",
        "required": [
          "elderly_user_id",
          "device_id"
        ],
        "title": "BindFamilyReq",
        "description": "家属绑定老人请求"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TokenResp": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "TokenResp",
        "description": "Token 响应"
      },
...
      "UserOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "group_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "mfa_enabled": {
            "type": "boolean",
            "title": "Mfa Enabled",
            "default": false
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role",
          "created_at"
        ],
        "title": "UserOut",
        "description": "用户信息响应"
      },
      "UserUpdate": {
        "properties": {
          "username": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Username",
            "description": "昵称"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone",
            "description": "手机号"
          }
        },
        "type": "object",
        "title": "UserUpdate",
        "description": "更新用户信息请求（昵称与手机号均可修改，二选一或全部）"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WebAuthnLoginIn": {
        "properties": {
          "credential": {
            "type": "object",
            "title": "Credential"
          },
          "challenge_token": {
            "type": "string",
            "title": "Challenge Token"
          }
        },
        "type": "object",
        "required": [
          "credential",
          "challenge_token"
        ],
        "title": "WebAuthnLoginIn"
      },
      "WebAuthnOptionsOut": {
        "properties": {
          "options": {
            "type": "object",
            "title": "Options"
          },
          "challenge_token": {
            "type": "string",
            "title": "Challenge Token"
          }
        },
        "type": "object",
        "required": [
          "options",
          "challenge_token"
        ],
        "title": "WebAuthnOptionsOut"
      },
      "WebAuthnRegisterIn": {
        "properties": {
          "credential": {
            "type": "object",
            "title": "Credential"
          },
          "challenge_token": {
            "type": "string",
            "title": "Challenge Token"
          },
          "nickname": {
            "type": "string",
            "title": "Nickname"
          }
        },
        "type": "object",
        "required": [
          "credential",
          "challenge_token"
        ],
        "title": "WebAuthnRegisterIn"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}